+++ engineFoxtrot/Engine/Fps/FrameData.h
@@ -1,10 +1,7 @@
 #pragma once
-#include <iostream>
 
 #define FRAME_VALUES 100
 #define TIMESTEP 1.f
-
-#undef main
 
 #if(EXPORT)
 class DLLEXPORT FrameData
@@ -15,7 +12,7 @@ class FrameData
 public:
 	FrameData();
 	~FrameData();
-	float calculateAverageFps();
+	double calculateAverageFps();
 	void startTimer();
 	static double gameFps;
 	static double renderFps;
@@ -23,5 +20,5 @@ public:
 private:
 	list<chrono::duration<double>> frametimes;
 	chrono::high_resolution_clock::time_point startTime;
-	double framespersecond;
+	double framesPerSecond = 0;
 };
