+++ engineFoxtrot/Engine/Structs/fpsStructs.cpp
@@ -1,0 +1,2 @@
+#include "stdafx.h"
+#include "fpsStructs.h"
