--- engineFoxtrot/Engine/Structs/structs.cpp
@@ -1,2 +1,0 @@
-#include "stdafx.h"
-#include "structs.h"
