+++ engineFoxtrot/Engine/Video/VideoEngine.h
@@ -1,15 +1,15 @@
 #pragma once
 #include "./VideoFacade.h"
 #include "../Fps/FrameData.h"
-#include "./Structs/structs.h"
+#include "./Structs/fpsStructs.h"
 
 #define NO_RED 0
 #define NO_BLUE 0
 #define NO_GREEN 0
 
-#define FPS_X 999
-#define Y_ZERO 0
-#define SDL_FPS_Y 36
+#define FPS_X_POSITION 999
+#define Y_POSITION_TOP_OF_SCREEN 0
+#define FPS_Y_POSITION_OFFSET 36
 
 #if(EXPORT)
 struct DLLEXPORT Sprite
@@ -53,9 +53,10 @@ public:
 
 	vector <Object*>* pointerToObjectVector = nullptr;
 
-	FrameData* frameData = nullptr;
 private:
 	IVideoFacade* videoFacade = new VideoFacade;
+
+	FrameData* frameData = nullptr;
 
 	bool shouldDrawFps = true; //TODO Should be set to false when toggle via button is added, see Todo above.
 };
