+++ engineFoxtrot/engineFoxtrot.vcxproj
@@ -185,7 +185,7 @@
     <ClCompile Include="Engine\Video\VideoFacade.cpp" />
     <ClCompile Include="Engine\Sound\SoundFacade.cpp" />
     <ClCompile Include="Engine\Sound\SoundEngine.cpp" />
-    <ClCompile Include="Engine\Structs\structs.cpp" />
+    <ClCompile Include="Engine\Structs\fpsStructs.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Engine\Events\AppTickEvent30.h" />
@@ -236,7 +236,7 @@
     <ClInclude Include="Engine\Sound\ISoundFacade.h" />
     <ClInclude Include="Engine\Sound\SoundFacade.h" />
     <ClInclude Include="Engine\Sound\SoundEngine.h" />
-    <ClInclude Include="Engine\Structs\structs.h" />
+    <ClInclude Include="Engine\Structs\fpsStructs.h" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="Engine\Documentation\PointerUsage.md" />
