+++ engineFoxtrot/engineFoxtrot.vcxproj.filters
@@ -25,7 +25,7 @@
     <ClCompile Include="Engine\Sound\SoundFacade.cpp" />
     <ClCompile Include="Engine\Sound\SoundEngine.cpp" />
     <ClCompile Include="Engine\Fps\FrameData.cpp" />
-    <ClCompile Include="Engine\Structs\structs.cpp" />
+    <ClCompile Include="Engine\Structs\fpsStructs.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Engine\Events\Event.h" />
@@ -76,7 +76,7 @@
     <ClInclude Include="Engine\Events\AppTickEvent60.h" />
     <ClInclude Include="Engine\Events\AppTickEvent30.h" />
     <ClInclude Include="Engine\Fps\FrameData.h" />
-    <ClInclude Include="Engine\Structs\structs.h" />
+    <ClInclude Include="Engine\Structs\fpsStructs.h" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="Engine\Documentation\PointerUsage.md" />
